Fix some issues found by Visual C++ warnings

+++ corvis/src/filter/filter.cpp
@@ -615,7 +615,7 @@ void filter_setup_next_frame(struct filter *f, const uint8_t *image, sensor_cloc
     //TODO: implement feature_single ?
 }
 
-static uint64_t get_raw_depth(const camera_data &cam, int x, int y)
+static uint16_t get_raw_depth(const camera_data &cam, int x, int y)
 {
     //TODO: make this more efficient if needed
     int dx = (x * cam.depth->width) / cam.width;
@@ -628,7 +628,7 @@ static uint64_t get_raw_depth(const camera_data &cam, int x, int y)
 static float get_depth_for_point(const camera_data &cam, int x, int y)
 {
     uint16_t depth_mm = get_raw_depth(cam, x, y);
-    return depth_mm / 1000.;
+    return depth_mm / 1000.0f;
 //    if(depth.image[y * depth.stride + x]) return depth.image[y * stride + x] / 1000.f;
 //    if(x == 0 || y == 0 || x == depth.width - 1 || y == depth.height - 1) return 0;
 }
